mohamad_email/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth[requests]
google-generativeai